+++ tests/testthat/test-mocktest.R
@@ -1,20 +1,24 @@
 
-skip_if_not_installed("mocktest", "0.1.0")
+pkg_dir <- system.file("testdata", "mocktest", package = "mockthat")
+pkg_nme <- pkgload::pkg_name(pkg_dir)
 
-Sys.setenv(TESTTHAT_PKG = "mocktest")
+pkgload::load_all(pkg_dir, attach_testthat = FALSE)
+
+withr::defer(pkgload::unload(pkg_nme))
+withr::local_envvar(TESTTHAT_PKG = pkg_nme)
 
 test_that("can change value of internal function", {
 
-  val <- mocktest:::test_non_exported()
+  val <- test_non_exported()
 
   res <- mockthat::with_mock(
     test_non_exported_nested = function() 5,
-    mocktest:::test_non_exported()
+    test_non_exported()
   )
 
   expect_equal(res, 5)
   expect_error(expect_equal(res, val), class = "expectation_failure")
-  expect_equal(mocktest:::test_non_exported(), val)
+  expect_equal(test_non_exported(), val)
 
   # and value is restored on error
   expect_error(
@@ -24,7 +28,7 @@ test_that("can change value of internal function", {
     )
   )
 
-  expect_equal(mocktest:::test_non_exported(), val)
+  expect_equal(test_non_exported(), val)
 })
 
 test_that("mocks can access local variables", {
@@ -33,7 +37,7 @@ test_that("mocks can access local variables", {
 
   mockthat::with_mock(
     test_non_exported_nested = function() x,
-    expect_equal(mocktest:::test_non_exported(), 5)
+    expect_equal(test_non_exported(), 5)
   )
 })
 
@@ -73,32 +77,46 @@ test_that("can't mock base functions", {
     ),
     "Can't mock functions in base package"
   )
+})
+
+test_that("can sometimes mock utils functions", {
+
+  res <- mockthat::with_mock(
+    show_struct = "bar",
+    test_utils("foo", "struct")
+  )
+
+  expect_identical(res, "bar")
 
   expect_error(
-    mockthat::with_mock(
-      head = function(...) tail(...), test_utils(mtcars)
-    ),
+    mockthat::with_mock(str = "bar", test_utils("foo", "struct")),
     "Can't mock functions in utils package"
   )
 
   expect_error(
-    mockthat::with_mock(
-      `utils::head` = function(...) utils::tail(...), test_utils(mtcars)
-    ),
+    mockthat::with_mock(`utils::str` = "bar", test_utils("foo", "struct")),
+    "Can't mock functions in utils package"
+  )
+
+  res <- mockthat::with_mock(
+    head = "bar",
+    test_utils("foo", "head")
+  )
+
+  expect_identical(res, "bar")
+
+  expect_error(
+    mockthat::with_mock(`utils::head` = "bar", test_utils("foo", "head")),
     "Can't mock functions in utils package"
   )
 
   expect_error(
-    mockthat::with_mock(
-      tail = function(...) head(...), test_utils(mtcars)
-    ),
+    mockthat::with_mock(tail = "bar", test_utils("foo", "tail")),
     "Can't mock functions in utils package"
   )
 
   expect_error(
-    mockthat::with_mock(
-      `utils::tail` = function(...) utils::head(...), test_utils(mtcars)
-    ),
+    mockthat::with_mock(`utils::tail` = "bar", test_utils("foo", "tail")),
     "Can't mock functions in utils package"
   )
 })
@@ -107,7 +125,7 @@ test_that("can't mock non-existing", {
 
   expect_error(
     mockthat::with_mock(..bogus.. = identity, TRUE),
-    "Function `\\.\\.bogus\\.\\.` not found in environment mocktest"
+    paste("Function `\\.\\.bogus\\.\\.` not found in environment", pkg_nme)
   )
 })
 
@@ -115,7 +133,7 @@ test_that("can't mock non-function", {
 
   expect_error(
     mockthat::with_mock(test_symbol = FALSE, TRUE),
-    "Function `test_symbol` not found in environment mocktest"
+    paste("Function `test_symbol` not found in environment", pkg_nme)
   )
 })
 
@@ -202,4 +220,95 @@ test_that(
   )
 
   expect_equal(x, 3)
+})
+
+test_that("mock within S3 dispatch", {
+
+  x <- structure(1L, class = "test_cls")
+
+  expect_identical(test_s3(x), 10L)
+  expect_identical(mockthat::with_mock(test_non_exported = 5L, test_s3(x)), 5L)
+
+  skip_if(getRversion() < "3.5.0")
+
+  expect_identical(test_s3(1L), 5L)
+  expect_identical(mockthat::with_mock(test_s3.default = 2L, test_s3(1L)), 2L)
+})
+
+test_that("mock depends", {
+
+  res <- mockthat::with_mock(
+    curl_dl_mem = "bar",
+    test_depends("foo", "mem")
+  )
+
+  expect_identical(res, "bar")
+
+  res <- mockthat::with_mock(
+    curl_fetch_disk = "bar",
+    test_depends("foo", "disk")
+  )
+
+  expect_identical(res, "bar")
+
+  res <- mockthat::with_mock(
+    `curl::curl_fetch_stream` = "bar",
+    test_depends("foo", "stream")
+  )
+
+  expect_identical(res, "bar")
+})
+
+test_that("mock imports", {
+
+  res <- mockthat::with_mock(
+    jsonlite_pretty = "bar",
+    test_imports("foo", "pretty")
+  )
+
+  expect_identical(res, "bar")
+
+  res <- mockthat::with_mock(
+    minify = "bar",
+    test_imports("foo", "mini")
+  )
+
+  expect_identical(res, "bar")
+
+  res <- mockthat::with_mock(
+    `jsonlite::fromJSON` = "bar",
+    test_imports("foo", "conv")
+  )
+
+  expect_identical(res, "bar")
+})
+
+test_that("mock returns", {
+
+  obj <- "bar"
+
+  res <- mockthat::with_mock(
+    curl_dl_mem = obj,
+    test_depends("foo", "mem")
+  )
+
+  expect_identical(res, obj)
+
+  obj <- c("bar", "baz")
+
+  res <- mockthat::with_mock(
+    curl_dl_mem = obj,
+    test_depends("foo", "mem")
+  )
+
+  expect_identical(res, obj)
+
+  obj <- list(a = 1, b = 2)
+
+  res <- mockthat::with_mock(
+    curl_dl_mem = obj,
+    test_depends("foo", "mem")
+  )
+
+  expect_identical(res, obj)
 })